Feat Recipe API (#5)
* Add Recipe model to core app

* Add Recipe API setup

Added the basic setup for the Recipe API, including AppConfig, url mappings, serializers, and views. This is necessary to provide the backend functionality for creating, updating, retrieving, and deleting recipes. The RecipeViewSet was added to manage the CRUD functionality for recipes.

* Refactor user API URLs and enhance test readability

* Add recipe model test

* Add tests for recipe APIs

This commit introduces a new file `test_recipe_api.py` in `app/recipe/tests/`. The file contains tests to validate the authenticated and unauthenticated access to recipe APIs. This includes testing for creating, retrieving, updating, and deleting recipes. Also, it includes checks for ensuring a user can only operate on his/her own recipes.

* Add 'recipe' to Django app settings and admin