Add Tag model and API related functionality
A new model 'Tag' has been introduced, along with related functionalities that include creating list of tags, limiting tags to authenticated user, etc. This model is linked to Recipe model as many-to-many relation. New migration file is created for this addition. Also, test cases have been developed to ensure the quality of the implemented features.