Add tag management in Recipe Serializer
Integrated the CRUD functionality for tags within the Recipe Serializer. A new TagSerializer is defined in the serializers.py module, and it is included in the fields of the Recipe serializer. Tags can be included while creating and updating recipes, and they're automatically added if they don't exist. Additionally, the TagViewSet in views.py is now capable of updating and destroying tags.