Feat Tags API (#6)
* Add Tag model and API related functionality

A new model 'Tag' has been introduced, along with related functionalities that include creating list of tags, limiting tags to authenticated user, etc. This model is linked to Recipe model as many-to-many relation. New migration file is created for this addition. Also, test cases have been developed to ensure the quality of the implemented features.

* Update URL routing and admin panel for recipe app

* Add tag management in Recipe Serializer

Integrated the CRUD functionality for tags within the Recipe Serializer. A new TagSerializer is defined in the serializers.py module, and it is included in the fields of the Recipe serializer. Tags can be included while creating and updating recipes, and they're automatically added if they don't exist. Additionally, the TagViewSet in views.py is now capable of updating and destroying tags.

* Add update and delete tests to tag API

A new functionality for updating and deleting tag has been added to the tag API, which was previously missing. Two new test cases: `test_update_tag` and `test_delete_tag` have been added to cover update and delete operations, validating both HTTP status and database changes.

* Added tests for tag handling in recipes

This commit extends the test coverage in the Recipe API by adding several unit tests that handle tags associated with recipes. This includes creating recipes with new and existing tags, updating recipes with new or existing tags, and testing the removal of tags from a recipe.