Feat Ingredients API (#7)
* "Add 'Ingredient' model and test

A new model 'Ingredient' was created with 'name' and 'user' fields. The model was then linked to the 'Recipe' model with 'ManyToMany' relationship. A new test for creating an ingredient was also added in 'test_models.py'."

* Add app as a source directory in recipe-app-api module

* Update version pinning in requirements.txt

* Add Ingredient model to admin registration

* Add Ingredient model to Recipe serializer

Added Ingredient model to Recipe serializer and created IngredientSerializer. Refactored the RecipeSerializer to accommodate new changes, and modified the instance creation and update methods, `_get_or_create_tags` to `_get_or_create_items`.

* Add tests for ingredients API

Added both public and private tests for the ingredients API to ensure proper functioning. The public test checks if authentication is necessary to retrieve ingredients. The private test checks if the user can retrieve a list of ingredients and also verifies that the list of ingredients is limited to the authenticated user. This increases the stability and reliability of the ingredients API.

* Refactor test utilities and extend recipe test cases

This modification refactors the test utility functions for creating users into a separate utils.py module. This change enhances the code reusability and makes the codebase cleaner and more maintainable. Additionally, new test cases were added for the recipe API. These tests cover creating a recipe with ingredients, creating a recipe with existing ingredients, creating an ingredient when updating a recipe, assigning an existing ingredient when updating a recipe, and clearing a recipe's ingredients. This improves the comprehensiveness of the test suite, ensuring that the related functionalities work as expected.

* Refactor recipe view for code efficiency

This commit involves the removal of duplicate code and unnecessary imports from the recipe view file. The change introduces a BaseViewSet that covers common functionality and streamlines the file. Now the RecipeViewSet, TagViewSet and new IngredientViewSet inherit from this BaseViewSet, thus making the code DRY and more maintainable. The IngredientViewSet was added to handle Ingredient operations in recipe API.

* Add IngredientViewSet to urls