Add image upload to RecipeViewSet
A new action was added to RecipeViewSet for uploading images related to a recipe. The user can now upload an image, which will be validated and saved. This change involved importing additional REST Framework modules to manage HTTP statuses and responses. If the upload is unsuccessful, an HTTP 400 status is returned along with the validation errors.