Feat Filtering by Tags and Ingredients (#9)
* Add Pillow to project requirements

* Add JPEG support and set up static/media folders

* Add volume for static data to docker-compose.yml

* Add image field to Recipe model and generate unique file path

* Add RecipeImageSerializer to serializers.py

* Update static and media URL settings

The static file serve URLs were updated along with the addition of media file serve URL in the settings.py. The changes were necessary to accommodate the refactored file structure and allow for static and media file handling. Also added settings for 'Spectacular Schema Generation' for OpenAPI documentation.

* Add static files URL handling for debug mode

This change adds static files URLs for handling media files in debug mode. This update enables serving media files during local development. The static file handling is added to the end of urlpatterns in urls.py file and only applies if the DEBUG setting is True.

* Add image upload to RecipeViewSet

A new action was added to RecipeViewSet for uploading images related to a recipe. The user can now upload an image, which will be validated and saved. This change involved importing additional REST Framework modules to manage HTTP statuses and responses. If the upload is unsuccessful, an HTTP 400 status is returned along with the validation errors.

* Add Image Upload Tests in recipe API

An ImageUploadTests class was added to `test_recipe_api.py` to handle the testing of new functionality: image upload. This includes test cases to upload an image to a recipe and to handle invalid image uploads. This is needed to ensure the image upload functionality works as expected and handles errors gracefully.

* Add UUID to image file path in recipe model tests

Added a mock UUID to the file path in the recipe model tests.

* Add image field to Recipe model

A new field 'image' is added to the Recipe model via a migration (0005_recipe_image.py). The 'image' field allows a user to upload an image for a recipe. The 'image' field is also included in the Recipe serializer, so it can be used in API requests.

* Refactor test utils into shared methods for better organization

Moved create_recipe, create_tag, and create_ingredient methods from app/recipe/tests/test_recipe_api.py to app/recipe/tests/utils.py, allowing for shared use across multiple test files. This change improves code organization, reduces redundancy, and allows for easier future modifications. Furthermore, additional tests have been added to filter recipes by ingredients and tags, thus enhancing current functionality.

* Add filtering to recipe API

Enhancements were made to the recipe API views to include more sophisticated filtering. The BaseRecipeAttrViewSet and RecipeViewSet classes now support filtering attributes such as 'assigned_only', 'tags', and 'ingredients'. This change provides users with more flexibility when retrieving recipes and attributes, specifically those assigned to recipes or tagged in a specific way. OpenAPI documentation was also updated to reflect these changes.