Add debug environment variable, and update Dockerfile
The `DEBUG` environment variable has been added to the application service in the docker-compose file for easier debugging during development.

In the Dockerfile, `scripts` directory is copied into the Docker image to aid application setup, and `linux-headers` included in the build dependencies. The `scripts` directory was added to the `PATH` environment variable, and an entry point command (`CMD ["run.sh"]`) executing a new script - `run.sh`, was added. The new script should make the initialization of Django easier.