Add Docker setup for application deployment
Added Dockerfile in proxy directory and docker-compose-deploy.yml to the root folder. The Dockerfile encapsulates the application within a Nginx server, and docker-compose file to manage the services required for deployment such as DB and the app itself. Setting up Docker will ensure smooth deployment and scaling of the application.