Add Deployment Configuration (#10)
* Update settings.py for security improvements

* Add debug environment variable, and update Dockerfile

The `DEBUG` environment variable has been added to the application service in the docker-compose file for easier debugging during development.

In the Dockerfile, `scripts` directory is copied into the Docker image to aid application setup, and `linux-headers` included in the build dependencies. The `scripts` directory was added to the `PATH` environment variable, and an entry point command (`CMD ["run.sh"]`) executing a new script - `run.sh`, was added. The new script should make the initialization of Django easier.

* Add Docker setup for application deployment

Added Dockerfile in proxy directory and docker-compose-deploy.yml to the root folder. The Dockerfile encapsulates the application within a Nginx server, and docker-compose file to manage the services required for deployment such as DB and the app itself. Setting up Docker will ensure smooth deployment and scaling of the application.

* Add run.sh scripts, sample .env file, and nginx configurations

New .sh scripts have been added (in scripts and proxy directories) to handle several automated startup tasks such as waiting for the database to start, running Django migrations and collecting static files. Added a sample .env file for setting up environment variables. Also, added nginx configuration files with a default config for serving static files and proxying to the application server. These changes are essential to set the application ready for deployment.

* Add uwsgi to requirements.txt