Update README and Workflow (#12)
* Update check workflow to run on main branch only

* Added newline at the end of urls.py

* Update README with project details

The README.md file has been revised to include a detailed description of the Recipe App API project. In addition to the project overview, the updated README now contains information about the features, technologies used, and instructions on building and running the application. This change aims to provide clearer project information for developers, users, and potential contributors.